Raspberry_SmartTemperature/requirements.txt
PyMySQL>=0.10.0
pyserial>=3.4
urllib3>=1.24.1
